pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sigilderg-finetuner"
version = "2.9.0"
description = "Model finetuner for the SigilDERG Ecosystem using QLoRA"
readme = "README.md"
requires-python = ">=3.12.10"
license = {text = "MIT"}
authors = [
    {name = "Dave Tofflemire"}
]
keywords = ["machine-learning", "llm", "fine-tuning", "qlora", "rust", "transformers"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "transformers>=4.44.0",
    "accelerate>=0.33.0",
    "datasets>=2.20.0",
    "bitsandbytes>=0.43.1",
    "trl>=0.9.6",
    "peft>=0.12.0",
    "evaluate>=0.4.2",
    "sentencepiece>=0.1.99",
    "tqdm>=4.66.0",
    "huggingface_hub>=0.24.0",
    "protobuf<5",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "jsonlines>=4.0.0",
    "pydantic>=2.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "tensorboard>=2.14.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
flash-attention = ["flash-attn>=2.5.6"]
pipeline = [
    "sigil-pipeline>=2.1.0",
]
evaluation = [
    "human-eval-rust>=2.1.0",
]
ecosystem = [
    "sigil-pipeline>=2.1.0",
    "human-eval-rust>=2.1.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "pytest-asyncio>=0.21.0",
    "hypothesis>=6.92.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "isort>=5.12.0",
    "flake8>=7.0.0",
    "mypy>=1.5.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "hypothesis>=6.92.0",
]

[project.scripts]
sigilderg-train = "rust_qlora.train:main"
sigilderg-eval = "rust_qlora.eval_rust:main"
sigilderg-sweep = "rust_qlora.hyperparameter_sweep:main"

[project.urls]
"Related Projects" = "https://github.com/Superuser666-Sigil"
"Pipeline" = "https://github.com/Superuser666-Sigil/SigilDERG-Data_Production"
"Finetuner" = "https://github.com/Superuser666-Sigil/SigilDERG-Finetuner"
"Evaluation" = "https://github.com/Superuser666-Sigil/human-eval-Rust"

[tool.setuptools]
packages = ["rust_qlora"]

[tool.setuptools.package-dir]
rust_qlora = "rust-qlora"

[tool.setuptools.package-data]
rust_qlora = ["configs/*.yml"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line length handled by black

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "gpu: marks tests that require GPU",
]

[tool.coverage.run]
source = ["rust_qlora"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.isort]
profile = "black"
line_length = 100

